refactor: move each inverter in a separate file
* remove redundant inverter mapping methods
* split up test data into multiple files
