+++ solax/inverters/x1_mini.py
@@ -1,0 +1,51 @@
+import voluptuous as vol
+from solax.inverter import InverterPost
+from solax.utils import startswith
+
+
+class X1Mini(InverterPost):
+    # pylint: disable=duplicate-code
+    _schema = vol.Schema({
+        vol.Required('type'): vol.All(
+            str,
+            startswith("X1-")
+        ),
+        vol.Required('SN'): str,
+        vol.Required('ver'): str,
+        vol.Required('Data'): vol.Schema(
+            vol.All(
+                [vol.Coerce(float)],
+                vol.Length(min=69, max=69),
+            )
+        ),
+        vol.Required('Information'): vol.Schema(
+            vol.All(
+                vol.Length(min=9, max=9)
+                )
+            ),
+    }, extra=vol.REMOVE_EXTRA)
+
+    _sensor_map = {
+        'PV1 Current':                (0, 'A'),
+        'PV2 Current':                (1, 'A'),
+        'PV1 Voltage':                (2, 'V'),
+        'PV2 Voltage':                (3, 'V'),
+
+        'Output Current':             (4, 'A'),
+        'Network Voltage':            (5, 'V'),
+        'AC Power':                   (6, 'W'),
+
+        'Inverter Temperature':       (7, 'C'),
+        'Today\'s Energy':            (8, 'kWh'),
+        'Total Energy':               (9, 'kWh'),
+        'Exported Power':             (10, 'W'),
+        'PV1 Power':                  (11, 'W'),
+        'PV2 Power':                  (12, 'W'),
+
+        'Total Feed-in Energy':       (41, 'kWh'),
+        'Total Consumption':          (42, 'kWh'),
+
+        'Power Now':                  (43, 'W'),
+        'Grid Frequency':             (50, 'Hz'),
+    }
+    # pylint: enable=duplicate-code
