+++ solax/inverters/x_hybrid.py
@@ -1,0 +1,78 @@
+import json
+import aiohttp
+import voluptuous as vol
+from solax.inverter import Inverter, InverterResponse
+
+
+class XHybrid(Inverter):
+    """
+    Tested with:
+    * SK-TL5000E
+    """
+    _schema = vol.Schema({
+        vol.Required('method'): str,
+        vol.Required('version'): str,
+        vol.Required('type'): str,
+        vol.Required('SN'): str,
+        vol.Required('Data'): vol.Schema(
+            vol.All(
+                [vol.Coerce(float)],
+                vol.Any(vol.Length(min=58, max=58), vol.Length(min=68, max=68))
+                )
+            ),
+        vol.Required('Status'): vol.All(vol.Coerce(int), vol.Range(min=0)),
+    }, extra=vol.REMOVE_EXTRA)
+
+    # key: name of sensor
+    # value.0: index
+    # value.1: unit (String) or None
+    # from https://github.com/GitHobi/solax/wiki/direct-data-retrieval
+    _sensor_map = {
+        'PV1 Current':                (0, 'A'),
+        'PV2 Current':                (1, 'A'),
+        'PV1 Voltage':                (2, 'V'),
+        'PV2 Voltage':                (3, 'V'),
+
+        'Output Current':             (4, 'A'),
+        'Network Voltage':            (5, 'V'),
+        'Power Now':                  (6, 'W'),
+
+        'Inverter Temperature':       (7, 'C'),
+        'Today\'s Energy':            (8, 'kWh'),
+        'Total Energy':               (9, 'kWh'),
+        'Exported Power':             (10, 'W'),
+        'PV1 Power':                  (11, 'W'),
+        'PV2 Power':                  (12, 'W'),
+
+        'Battery Voltage':            (13, 'V'),
+        'Battery Current':            (14, 'A'),
+        'Battery Power':              (15, 'W'),
+        'Battery Temperature':        (16, 'C'),
+        'Battery Remaining Capacity': (17, '%'),
+
+        'Month\'s Energy':            (19, 'kWh'),
+
+        'Grid Frequency':             (50, 'Hz'),
+        'EPS Voltage':                (53, 'V'),
+        'EPS Current':                (54, 'A'),
+        'EPS Power':                  (55, 'W'),
+        'EPS Frequency':              (56, 'Hz'),
+    }
+
+    @classmethod
+    async def make_request(cls, host, port=80, pwd='', headers=None):
+        base = 'http://{}:{}/api/realTimeData.htm'
+        url = base.format(host, port)
+        async with aiohttp.ClientSession() as session:
+            async with session.get(url) as req:
+                garbage = await req.read()
+        formatted = garbage.decode("utf-8")
+        formatted = formatted.replace(",,", ",0.0,").replace(",,", ",0.0,")
+        json_response = json.loads(formatted)
+        response = cls.schema()(json_response)
+        return InverterResponse(
+            data=cls.map_response(response['Data']),
+            serial_number=response['SN'],
+            version=response['version'],
+            type=response['type']
+        )
