+++ tests/samples/expected_values.py
@@ -1,0 +1,380 @@
+XHYBRID_VALUES = {
+    "Today's Energy": 8.0,
+    "Battery Current": 14.0,
+    "Month's Energy": 19.0,
+    "Battery Power": 15,
+    "Battery Remaining Capacity": 17,
+    "Battery Temperature": 16,
+    "Battery Voltage": 13,
+    "EPS Current": 54,
+    "EPS Frequency": 56,
+    "EPS Power": 55,
+    "EPS Voltage": 53,
+    "Exported Power": 10,
+    "Grid Frequency": 50,
+    "Inverter Temperature": 7,
+    "Network Voltage": 5,
+    "Output Current": 4,
+    "PV1 Current": 0,
+    "PV1 Power": 11,
+    "PV1 Voltage": 2,
+    "PV2 Current": 1,
+    "PV2 Power": 12,
+    "PV2 Voltage": 3,
+    "Power Now": 6,
+    "Total Energy": 9,
+}
+
+
+X3_VALUES = {
+    "PV1 Current": 0,
+    "PV2 Current": 1,
+    "PV1 Voltage": 2,
+    "PV2 Voltage": 3,
+    "Output Current Phase 1": 4,
+    "Network Voltage Phase 1": 5,
+    "AC Power": 6,
+    "Inverter Temperature": 7,
+    "Today's Energy": 8,
+    "Total Energy": 9,
+    "Exported Power": 10,
+    "PV1 Power": 11,
+    "PV2 Power": 12,
+    "Battery Voltage": 13,
+    "Battery Current": 14,
+    "Battery Power": 15,
+    "Battery Temperature": 16,
+    "Battery Remaining Capacity": 21,
+    "Total Feed-in Energy": 41,
+    "Total Consumption": 42,
+    "Output Current Phase 2": 46,
+    "Output Current Phase 3": 47,
+    "Network Voltage Phase 2": 48,
+    "Network Voltage Phase 3": 49,
+    "Grid Frequency Phase 1": 50,
+    "Grid Frequency Phase 2": 51,
+    "Grid Frequency Phase 3": 52,
+    "Power Now Phase 1": 43,
+    "Power Now Phase 2": 44,
+    "Power Now Phase 3": 45,
+    "EPS Voltage": 53,
+    "EPS Current": 54,
+    "EPS Power": 55,
+    "EPS Frequency": 56,
+}
+
+X3_HYBRID_VALUES = {
+    "PV1 Current": 0,
+    "PV2 Current": 0,
+    "PV1 Voltage": 0,
+    "PV2 Voltage": 0,
+    "Output Current Phase 1": 0.9,
+    "Network Voltage Phase 1": 234,
+    "AC Power": 3189,
+    "Inverter Temperature": 42,
+    "Today's Energy": 15.2,
+    "Total Energy": 27,
+    "Exported Power": -25,
+    "PV1 Power": 0,
+    "PV2 Power": 0,
+    "Battery Voltage": 210.3,
+    "Battery Current": -15.7,
+    "Battery Power": -3321,
+    "Battery Temperature": 24,
+    "Battery Remaining Capacity": 68,
+    "Total Feed-in Energy": 2.03,
+    "Total Consumption": 23.41,
+    "Output Current Phase 2": 5.4,
+    "Output Current Phase 3": 6.9,
+    "Network Voltage Phase 2": 250,
+    "Network Voltage Phase 3": 251.9,
+    "Grid Frequency Phase 1": 50.01,
+    "Grid Frequency Phase 2": 50.01,
+    "Grid Frequency Phase 3": 50.01,
+    "Power Now Phase 1": 123,
+    "Power Now Phase 2": 1344,
+    "Power Now Phase 3": 1722,
+    "EPS Voltage": 0,
+    "EPS Current": 0,
+    "EPS Power": 0,
+    "EPS Frequency": 0,
+}
+
+X3V34_HYBRID_VALUES = {
+    "Network Voltage Phase 1": 246.8,
+    "Network Voltage Phase 2": 249,
+    "Network Voltage Phase 3": 250.8,
+    "Output Current Phase 1": 1.3,
+    "Output Current Phase 2": 1.4,
+    "Output Current Phase 3": 1,
+    "Power Now Phase 1": 266,
+    "Power Now Phase 2": 284,
+    "Power Now Phase 3": 136,
+    "PV1 Voltage": 537.7,
+    "PV2 Voltage": 463,
+    "PV1 Current": 1.7,
+    "PV2 Current": 0,
+    "PV1 Power": 958,
+    "PV2 Power": 0,
+    "Total PV Energy": 1731.9,
+    "Total PV Energy Resets": 0,
+    "Today's PV Energy": 16.4,
+    "Grid Frequency Phase 1": 50.03,
+    "Grid Frequency Phase 2": 50.03,
+    "Grid Frequency Phase 3": 50.03,
+    "Total Energy": 1483.3,
+    "Total Energy Resets": 0,
+    "Today's Energy": 10.3,
+    "Battery Voltage": 229.3,
+    "Battery Current": 0.9,
+    "Battery Power": 229,
+    "Battery Temperature": 22,
+    "Battery Remaining Capacity": 99,
+    "Total Battery Discharge Energy": 706.2,
+    "Total Battery Discharge Energy Resets": 0,
+    "Today's Battery Discharge Energy": 4.3,
+    "Battery Remaining Energy": 12.5,
+    "Total Battery Charge Energy": 814.2,
+    "Total Battery Charge Energy Resets": 0,
+    "Today's Battery Charge Energy": 9.1,
+    "Exported Power": -55,
+    "Total Feed-in Energy": 173.72,
+    "Total Feed-in Energy Resets": 0,
+    "Total Consumption": 598.77,
+    "Total Consumption Resets": 0,
+    "AC Power": 686,
+    "EPS Frequency": 0,
+    "EPS Total Energy": 0.6,
+    "EPS Total Energy Resets": 0,
+}
+
+X3V34_HYBRID_VALUES_NEGATIVE_POWER = {
+    "Network Voltage Phase 1": 236.4,
+    "Network Voltage Phase 2": 243.1,
+    "Network Voltage Phase 3": 238.6,
+    "Output Current Phase 1": -7.2,
+    "Output Current Phase 2": -7.1,
+    "Output Current Phase 3": -7.1,
+    "Power Now Phase 1": -1737,
+    "Power Now Phase 2": -1766,
+    "Power Now Phase 3": -1728,
+    "PV1 Voltage": 0,
+    "PV2 Voltage": 0,
+    "PV1 Current": 0,
+    "PV2 Current": 0,
+    "PV1 Power": 0,
+    "PV2 Power": 0,
+    "Total PV Energy": 9615.1,
+    "Total PV Energy Resets": 1,
+    "Today's PV Energy": 9.8,
+    "Grid Frequency Phase 1": 49.98,
+    "Grid Frequency Phase 2": 49.98,
+    "Grid Frequency Phase 3": 49.98,
+    "Total Energy": 8478.6,
+    "Total Energy Resets": 1,
+    "Today's Energy": 8.4,
+    "Battery Voltage": 204.6,
+    "Battery Current": 25,
+    "Battery Power": 5117,
+    "Battery Temperature": 24,
+    "Battery Remaining Capacity": 20,
+    "Total Battery Discharge Energy": 2469.6,
+    "Total Battery Discharge Energy Resets": 0,
+    "Today's Battery Discharge Energy": 4,
+    "Battery Remaining Energy": 2.6,
+    "Total Battery Charge Energy": 2887.6,
+    "Total Battery Charge Energy Resets": 0,
+    "Today's Battery Charge Energy": 6,
+    "Exported Power": -5743,
+    "Total Feed-in Energy": 3593.49,
+    "Total Feed-in Energy Resets": 5,
+    "Total Consumption": 945.62,
+    "Total Consumption Resets": 1,
+    "AC Power": -5233,
+    "EPS Frequency": 0,
+    "EPS Total Energy": 2.1,
+    "EPS Total Energy Resets": 0,
+}
+
+X3V34_HYBRID_VALUES_EPS_MODE = {
+    "Network Voltage Phase 1": 0,
+    "Network Voltage Phase 2": 0,
+    "Network Voltage Phase 3": 0,
+    "Output Current Phase 1": 0,
+    "Output Current Phase 2": 0,
+    "Output Current Phase 3": 0,
+    "Power Now Phase 1": 0,
+    "Power Now Phase 2": 0,
+    "Power Now Phase 3": 0,
+    "PV1 Voltage": 486.4,
+    "PV2 Voltage": 446.6,
+    "PV1 Current": 5.5,
+    "PV2 Current": 0,
+    "PV1 Power": 2678,
+    "PV2 Power": 0,
+    "Total PV Energy": 8378.7,
+    "Total PV Energy Resets": 1,
+    "Today's PV Energy": 26,
+    "Grid Frequency Phase 1": 0,
+    "Grid Frequency Phase 2": 0,
+    "Grid Frequency Phase 3": 0,
+    "Total Energy": 7387.4,
+    "Total Energy Resets": 1,
+    "Today's Energy": 17.8,
+    "Battery Voltage": 228.6,
+    "Battery Current": -0.99,
+    "Battery Power": -241,
+    "Battery Temperature": 27,
+    "Battery Remaining Capacity": 98,
+    "Total Battery Discharge Energy": 2124,
+    "Total Battery Discharge Energy Resets": 0,
+    "Today's Battery Discharge Energy": 6.1,
+    "Battery Remaining Energy": 12.4,
+    "Total Battery Charge Energy": 2501.1,
+    "Total Battery Charge Energy Resets": 0,
+    "Today's Battery Charge Energy": 11.9,
+    "Exported Power": 0,
+    "Total Feed-in Energy": 3174.79,
+    "Total Feed-in Energy Resets": 4,
+    "Total Consumption": 844.57,
+    "Total Consumption Resets": 1,
+    "AC Power": 0,
+    "EPS Frequency": 50,
+    "EPS Total Energy": 1.7,
+    "EPS Total Energy Resets": 0,
+}
+
+X1_VALUES = {
+    "PV1 Current": 0,
+    "PV2 Current": 1,
+    "PV1 Voltage": 2,
+    "PV2 Voltage": 3,
+    "Output Current": 4,
+    "Network Voltage": 5,
+    "AC Power": 6,
+    "Inverter Temperature": 7,
+    "Today's Energy": 8,
+    "Total Energy": 9,
+    "Exported Power": 10,
+    "PV1 Power": 11,
+    "PV2 Power": 12,
+    "Battery Voltage": 13,
+    "Battery Current": 14,
+    "Battery Power": 15,
+    "Battery Temperature": 16,
+    "Battery Remaining Capacity": 21,
+    "Total Feed-in Energy": 41,
+    "Total Consumption": 42,
+    "Grid Frequency": 50,
+    "Power Now": 43,
+    "EPS Voltage": 53,
+    "EPS Current": 54,
+    "EPS Power": 55,
+    "EPS Frequency": 56,
+}
+
+X1_MINI_VALUES = {
+    "PV1 Current": 0,
+    "PV2 Current": 1,
+    "PV1 Voltage": 2,
+    "PV2 Voltage": 3,
+    "Output Current": 4,
+    "Network Voltage": 5,
+    "AC Power": 6,
+    "Inverter Temperature": 7,
+    "Today's Energy": 8,
+    "Total Energy": 9,
+    "Exported Power": 10,
+    "PV1 Power": 11,
+    "PV2 Power": 12,
+    "Total Feed-in Energy": 41,
+    "Total Consumption": 42,
+    "Power Now": 43,
+    "Grid Frequency": 50,
+}
+
+X1_MINI_VALUES_V34 = {
+    "Network Voltage": 231.0,
+    "Output Current": 1.5,
+    "AC Power": 349,
+    "PV1 Voltage": 60.9,
+    "PV2 Voltage": 0,
+    "PV1 Current": 5.8,
+    "PV2 Current": 0,
+    "PV1 Power": 359,
+    "PV2 Power": 0,
+    "Grid Frequency": 50.0,
+    "Total Energy": 347.4,
+    "Today's Energy": 1.2,
+    "Total Feed-in Energy": 251.8,
+    "Total Consumption": 0,
+    "Power Now": 0,
+}
+
+X1_SMART_VALUES = {
+    "Network Voltage": 239.6,
+    "Output Current": 12.6,
+    "AC Power": 2956,
+    "PV1 Voltage": 432.9,
+    "PV2 Voltage": 233.8,
+    "PV1 Current": 4.1,
+    "PV2 Current": 5.7,
+    "PV1 Power": 1777,
+    "PV2 Power": 1336,
+    "Grid Frequency": 60.21,
+    "Total Energy": 121.5,
+    "Today's Energy": 13.7,
+    "Inverter Temperature": 43,
+    "Exported Power": 10,
+    "Total Feed-in Energy": 0.87,
+    "Total Consumption": 81.84,
+}
+
+QVOLTHYBG33P_VALUES = {
+    "Network Voltage Phase 1": 221.4,
+    "Network Voltage Phase 2": 223.8,
+    "Network Voltage Phase 3": 225.1,
+    "Output Current Phase 1": 1.1,
+    "Output Current Phase 2": 1.0,
+    "Output Current Phase 3": 1.2,
+    "Power Now Phase 1": 162.0,
+    "Power Now Phase 2": 136.0,
+    "Power Now Phase 3": 146.0,
+    "AC Power": 444.0,
+    "PV1 Voltage": 566.2,
+    "PV2 Voltage": 568.2,
+    "PV1 Current": 1.8,
+    "PV2 Current": 1.7,
+    "PV1 Power": 1050.0,
+    "PV2 Power": 977.0,
+    "Grid Frequency Phase 1": 50.02,
+    "Grid Frequency Phase 2": 50.01,
+    "Grid Frequency Phase 3": 50.02,
+    "Inverter Operation mode": "Normal",
+    "Exported Power": -6.0,
+    "Battery Voltage": 323.4,
+    "Battery Current": 5.0,
+    "Battery Power": 1616.0,
+    "Power Now": 451.0,
+    "Total Energy": 219.0,
+    "Total Energy Resets": 0.0,
+    "Total Battery Discharge Energy": 73.8,
+    "Total Battery Discharge Energy Resets": 0.0,
+    "Total Battery Charge Energy": 90.4,
+    "Total Battery Charge Energy Resets": 0.0,
+    "Today's Battery Discharge Energy": 0.0,
+    "Today's Battery Charge Energy": 8.1,
+    "Total PV Energy": 231.6,
+    "Total PV Energy Resets": 0.0,
+    "Today's Energy": 11.8,
+    "Total Feed-in Energy": 107.94,
+    "Total Feed-in Energy Resets": 0.0,
+    "Total Consumption": 145.44,
+    "Total Consumption Resets": 0.0,
+    "Today's Feed-in Energy": 1.66,
+    "Today's Consumption": 4.66,
+    "Battery Remaining Capacity": 95.0,
+    "Battery Temperature": 35.0,
+    "Battery Remaining Energy": 8.8,
+    "Battery Operation mode": "Self Use Mode",
+}
