+++ tests/samples/responses.py
@@ -1,0 +1,272 @@
+
+XHYBRID_DE01_RESPONSE = {
+    'method': 'uploadsn',
+    'version': 'Solax_SI_CH_2nd_20160912_DE02',
+    'type': 'AL_SE',
+    'SN': 'XXXXXXX',
+    'Data': [0, 1, 2, 3, 4, 5, 6, 7, 8, 9,
+             10, 11, 12, 13, 14, 15, 16, 17, 18, 19,
+             20, 21, 22, 23, 24, 25, 26, 27, 28, 29,
+             30, 31, 32, 33, 34, 35, 36, 37, 38, 39,
+             40, 41, 42, 43, 44, 45, 46, 47, 48, 49,
+             50, 51, 52, 53, 54, 55, 56, 57],
+    'Status': 2
+}
+
+XHYBRID_DE02_RESPONSE = {
+    'method': 'uploadsn',
+    'version': 'Solax_SI_CH_2nd_20160912_DE02',
+    'type': 'AL_SE',
+    'SN': 'XXXXXXX',
+    'Data': [0, 1, 2, 3, 4, 5, 6, 7, 8, 9,
+             10, 11, 12, 13, 14, 15, 16, 17, 18, 19,
+             20, 21, 22, 23, 24, 25, 26, 27, 28, 29,
+             30, 31, 32, 33, 34, 35, 36, 37, 38, 39,
+             40, 41, 42, 43, 44, 45, 46, 47, 48, 49,
+             50, 51, 52, 53, 54, 55, 56, 57, 58, 59,
+             60, 61, 62, 63, 64, 65, 66, 67],
+    'Status': 2
+}
+
+X1_BOOST_AIR_MINI_RESPONSE = {
+    "type": "X1-Boost-Air-Mini",
+    "SN": "XXXXXXX",
+    "ver": "2.32.6",
+    "Data": [0, 1, 2, 3, 4, 5, 6, 7, 8, 9,
+             10, 11, 12, 13, 14, 15, 16, 17, 18, 19,
+             20, 21, 22, 23, 24, 25, 26, 27, 28, 29,
+             30, 31, 32, 33, 34, 35, 36, 37, 38, 39,
+             40, 41, 42, 43, 44, 45, 46, 47, 48, 49,
+             50, 51, 52, 53, 54, 55, 56, 57, 58, 59,
+             60, 61, 62, 63, 64, 65, 66, 67, 68],
+    "Information": [2.500, 4, "X1-Boost-Air-Mini", "XXXXXXX",
+                    1, 3.25, 1.09, 1.10, 0.00]
+}
+
+X1_MINI_RESPONSE_V34 = {
+    "sn": "XXXXXXXXXX",
+    "ver": "2.034.06",
+    "type": 4,
+    "Data": [2310, 15, 349, 609, 0, 58, 0, 359, 0, 5000,
+             2, 3474, 0, 12, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+             0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 39, 0, 2518, 0, 0, 0, 0, 0, 0,
+             0, 0, 0, 0, 0, 0, 0, 41, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+             0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+             0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+             0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+             0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+             0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+             0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+             0, 0, 0, 0, 0],
+    "Information": [0.700, 4, "XXXXXXXXXXXXXX",
+                    1, 1.19, 0.00, 1.32, 0.00, 0.00, 1]
+}
+
+X1_SMART_RESPONSE = {
+    "sn": "XXXXXXX",
+    "ver": "2.033.20",
+    "type": 8,
+    "Data": [2396, 126, 2956, 4329, 2338, 41, 57, 1777, 1336, 6021,
+             2, 1215, 0, 137, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+             0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 43, 0, 80, 0, 0, 0, 0, 0, 0,
+             10, 65535, 87, 0, 8184, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+             0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+             0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+             0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+             0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+             0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+             0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+             0, 0, 0, 0, 0, 0, 0, 0, 0],
+    "Information": [8.0, 8, "XXXXXXX", 1, 1.07, 1.0, 1.05, 0.0]
+}
+
+X3_MIC_RESPONSE = {
+    "type": "X3-MIC",
+    "SN": "XXXXXXX",
+    "ver": "2.033.20",
+    "Data": [0, 1, 2, 3, 4, 5, 6, 7, 8, 9,
+             10, 11, 12, 13, 14, 15, 16, 17, 18, 19,
+             20, 21, 22, 23, 24, 25, 26, 27, 28, 29,
+             30, 31, 32, 33, 34, 35, 36, 37, 38, 39,
+             40, 41, 42, 43, 44, 45, 46, 47, 48, 49,
+             50, 51, 52, 53, 54, 55, 56, 57, 58, 59,
+             60, 61, 62, 63, 64, 65, 66, 67, 68, 69,
+             70, 71, 72, 73, 74, 75, 76, 77, 78, 79,
+             80, 81, 82, 83, 84, 85, 86, 87, 88, 89,
+             90, 91, 92, 93, 94, 95, 96, 97, 98, 99,
+             100, 101, 102],
+    "Information": [8.000, 7, "X3-MIC", "XXXXXXX",
+                    1, 1.10, 1.02, 1.09, 1.02]
+}
+
+X1_HYBRID_G3_RESPONSE = {
+    "type": "X1-Hybiyd-G3",
+    "SN": "XXXXXXX",
+    "ver": "2.033.20",
+    "Data": [0, 1, 2, 3, 4, 5, 6, 7, 8, 9,
+             10, 11, 12, 13, 14, 15, 16, 17, 18, 19,
+             20, 21, 22, 23, 24, 25, 26, 27, 28, 29,
+             30, 31, 32, 33, 34, 35, 36, 37, 38, 39,
+             40, 41, 42, 43, 44, 45, 46, 47, 48, 49,
+             50, 51, 52, 53, 54, 55, 56, 57, 58, 59,
+             60, 61, 62, 63, 64, 65, 66, 67, 68, 69,
+             70, 71, 72, 73, 74, 75, 76, 77, 78, 79,
+             80, 81, 82, 83, 84, 85, 86, 87, 88, 89,
+             90, 91, 92, 93, 94, 95, 96, 97, 98, 99,
+             100, 101],
+    "Information": [8.000, 7, "X3-MIC", "XXXXXXX",
+                    1, 1.10, 1.02, 1.09, 1.02]
+}
+
+X1_HYBRID_G3_2X_MPPT_RESPONSE = {
+    "type": "X1-Hybiyd-G3",
+    "SN": "XXXXXXXXXX",
+    "ver": "2.033.20",
+    "Data": [0, 1, 2, 3, 4, 5, 6, 7, 8, 9,
+             10, 11, 12, 13, 14, 15, 16, 17, 18, 19,
+             20, 21, 22, 23, 24, 25, 26, 27, 28, 29,
+             30, 31, 32, 33, 34, 35, 36, 37, 38, 39,
+             40, 41, 42, 43, 44, 45, 46, 47, 48, 49,
+             50, 51, 52, 53, 54, 55, 56, 57, 58, 59,
+             60, 61, 62, 63, 64, 65, 66, 67, 68, 69,
+             70, 71, 72, 73, 74, 75, 76, 77, 78, 79,
+             80, 81, 82, 83, 84, 85, 86, 87, 88, 89,
+             90, 91, 92, 93, 94, 95, 96, 97, 98, 99,
+             100, 101, 102, 103, 104, 105, 106],
+    "Information": [3.000, 3, "X1-Hybiyd-G3", "YYYYYYYYYYYYYY",
+                    1, 3.11, 0.00, 3.13, 1.05],
+    "battery": {
+        "brand": "0",
+        "masterVer": "0.00",
+        "slaveNum": "0",
+        "slaveVer": [0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00, 0.00],
+        "slaveType": [0, 0, 0, 0, 0, 0, 0, 0]
+    }
+}
+
+X3_HYBRID_G3_RESPONSE = {
+    "type": "X3-Hybiyd-G3",
+    "SN": "XXXXXXX",
+    "ver": "2.32.6",
+    "Data": [0, 1, 2, 3, 4, 5, 6, 7, 8, 9,
+             10, 11, 12, 13, 14, 15, 16, 17, 18, 19,
+             20, 21, 22, 23, 24, 25, 26, 27, 28, 29,
+             30, 31, 32, 33, 34, 35, 36, 37, 38, 39,
+             40, 41, 42, 43, 44, 45, 46, 47, 48, 49,
+             50, 51, 52, 53, 54, 55, 56, 57, 58, 59,
+             60, 61, 62, 63, 64, 65, 66, 67, 68, 69,
+             70, 71, 72, 73, 74, 75, 76, 77, 78, 79,
+             80, 81, 82, 83, 84, 85, 86, 87, 88, 89,
+             90, 91, 92, 93, 94, 95, 96, 97, 98, 99,
+             100, 101],
+    "Information": [0.000, 5, "X3-Hybiyd-G3", "XXXXXXX",
+                    1, 3.00, 0.00, 3.17, 1.01]
+}
+
+
+X3_HYBRID_G3_2X_MPPT_RESPONSE = {
+    "type": "X3-Hybiyd-G3",
+    "SN": "XXXXXXXXXX",
+    "ver": "2.033.20",
+    "Data": [0.0, 0.0, 0.0, 0.0, 0.9, 234.0, 3189, 42, 15.2, 27.0, -25, 0, 0,
+             210.30, -15.70, -3321, 24, 8.6, 0, 11.0, 1, 68, 232.4, 170.0,
+             31.0, 35.0, 22.6, 20.7, 3.8, 3.8, 2, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+             2.03, 23.41, 123, 1344, 1722, 5.4, 6.9, 250.0, 251.9, 50.01,
+             50.01, 50.01, 0.0, 0.0, 0, 0.00, 0, 0, 0, 0.00, 0, 0, 0, 0, 0.00,
+             0, 0, 2, 1, 26, 1.00, 0, 100, 10, 25.00, 25.00, 0, 0, 0, 0, 0.0,
+             10.8, 24.4, 0.0, 0, 2, 0.0, 0.0, 0, 0.0, 0.0, 0, 0, 0, 0, 1, 1,
+             0, 0, 0.00, 0.00, 1, 273, 212.3, -16.2, -3439],
+    "Information": [8.000, 5, "X3-Hybiyd-G3", "XXXXXXXX", 1, 4.47, 0.00, 4.34,
+                    1.05],
+    "battery": {
+        "brand": "83",
+        "masterVer": "1.11",
+        "slaveNum": "4",
+        "slaveVer": [1.13, 1.13, 1.13, 1.13]
+    }
+}
+
+X3_HYBRID_G3_2X_MPPT_RESPONSE_V34 = {
+    "type": 5,
+    "sn": "XXXXXXXXXX",
+    "ver": "2.034.06",
+    "Data": [2468, 2490, 2508, 13, 14, 10, 266, 284, 136, 5377, 4630, 17, 0,
+             958, 0, 5003, 5003, 5003, 2, 14833, 0, 103, 0, 0, 22930, 90, 229,
+             22, 99, 0, 7062, 0, 125, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+             100, 0, 41, 7777, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 65480,
+             0, 17372, 0, 59877, 0, 665, 41, 256, 2352, 1568, 20, 350, 202,
+             190, 41, 41, 81, 1, 1, 0, 0, 8142, 0, 17319, 0, 6, 0, 64851,
+             65535, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 83, 0, 55, 0, 0, 6, 0, 164,
+             43, 91, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 256, 12579, 783, 5381,
+             1107, 512, 8224, 8224, 0, 0, 4369, 0, 273, 2295, 5, 114, 4112,
+             4096, 25912, 31, 21302, 19778, 18003, 12355, 16697, 12354, 14132,
+             21302, 13110, 12338, 12337, 14386, 12354, 12852, 21302, 13110,
+             12338, 12337, 14386, 12354, 12340, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+             0, 0, 0, 0, 686, 1, 257, 257, 1794, 1025, 0, 22930, 0, 0, 0, 0, 0,
+             0, 0, 0, 0, 0, 0],
+    "Information": [8.000, 5, "XXXXXXXX", 1, 4.47, 0.00, 4.34, 1.05, 0.0, 1]
+}
+
+X3_HYBRID_G3_2X_MPPT_RESPONSE_V34_NEGATIVE_POWER = {
+    "type": 5,
+    "sn": "XXXXXXXXXX",
+    "ver": "2.034.06",
+    "Data": [2364, 2431, 2386, 65463, 65464, 65464, 63798, 63769, 63807, 0, 0,
+             0, 0, 0, 0, 4998, 4998, 4998, 2, 19251, 1, 84, 0, 0, 20460, 2500,
+             5117, 24, 20, 0, 24696, 0, 26, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+             0, 0, 100, 0, 36, 7577, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+             59792, 65535, 31674, 5, 29027, 1, 510, 36, 256, 2352, 1568, 310,
+             350, 217, 205, 36, 36, 258, 1, 1, 13, 0, 28876, 0, 30616, 1, 21,
+             0, 5236, 0, 65535, 65535, 0, 0, 0, 0, 0, 0, 0, 0, 27, 0, 936, 0,
+             0, 21, 0, 98, 40, 60, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 256, 4392,
+             7702, 5386, 1107, 512, 8481, 8481, 0, 0, 4369, 0, 273, 2037, 243,
+             4949, 3645, 3638, 12, 100, 21302, 19778, 18003, 12355, 16697,
+             12354, 14132, 21302, 13110, 12338, 12337, 14386, 12354, 12852,
+             21302, 13110, 12338, 12337, 14386, 12354, 12340, 0, 0, 0, 0, 0, 0,
+             0, 0, 0, 0, 0, 0, 0, 0, 60302, 1, 257, 257, 770, 1028, 0, 20460,
+             0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+    "Information": [8.000, 5, "XXXXXXXX", 1, 4.60, 0.00, 4.42, 1.05, 0.00, 1]
+}
+
+X3_HYBRID_G3_2X_MPPT_RESPONSE_V34_EPS_MODE = {
+    "type": 5,
+    "sn": "XXXXXXXXXX",
+    "ver": "2.034.06",
+    "Data": [0, 0, 0, 0, 0, 0, 0, 0, 0, 4864, 4466, 55, 0, 2678, 0, 0, 0, 0, 7,
+             8339, 1, 178, 0, 0, 22860, 65436, 65294, 27, 98, 0, 21240, 0, 124,
+             0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 100, 0, 52, 7541, 226,
+             2453, 205, 2223, 2276, 2426, 10, 107, 8, 172, 2441, 147, 5000, 0,
+             0, 0, 55339, 4, 18922, 1, 0, 52, 256, 2352, 1568, 0, 350, 275,
+             262, 41, 41, 225, 1, 1, 0, 0, 25011, 0, 18252, 1, 17, 0, 0, 0, 0,
+             0, 0, 0, 0, 0, 0, 0, 0, 0, 768, 0, 54, 0, 2, 17, 0, 260, 61, 119,
+             0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 256, 7971, 6926, 5385, 1107, 512,
+             8224, 8224, 0, 0, 4369, 0, 273, 2288, 65523, 65239, 4099, 4083,
+             13728, 87, 21302, 19778, 18003, 12355, 16697, 12354, 14132, 21302,
+             13110, 12338, 12337, 14386, 12354, 12852, 21302, 13110, 12338,
+             12337, 14386, 12354, 12340, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+             0, 0, 1, 257, 257, 1281, 1025, 0, 22860, 0, 0, 0, 0, 0, 0, 0, 0,
+             0, 0, 0],
+    "Information": [8.000, 5, "XXXXXXXX", 1, 4.60, 0.00, 4.42, 1.05, 0.00, 1]
+}
+
+QVOLTHYBG33P_RESPONSE_V34 = {
+    "sn": "SWX***",
+    "ver": "2.034.06",
+    "type": 14,
+    "Data": [
+        2214, 2238, 2251, 11, 10, 12, 162, 136, 146, 444, 5662, 5682, 18, 17,
+        1050, 977, 5002, 5001, 5002, 2, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        1, 65529, 65535, 0, 0, 0, 32340, 500, 1616, 3236, 50, 1618, 1, 50, 451,
+        256, 3841, 4875, 5634, 100, 0, 45, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 2190, 0, 36, 1, 0, 1, 738, 0, 904, 0, 0, 81, 2316, 0, 118, 0, 0, 0,
+        10794, 0, 14544, 0, 166, 0, 466, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 95, 1,
+        35, 88, 256, 3504, 2400, 115, 300, 352, 325, 34, 34, 8, 1620, 773,
+        12850, 12850, 12850, 0, 0, 0, 3389, 3384, 33876, 2, 20564, 12339,
+        18497, 12599, 18743, 12356, 14386, 20564, 12339, 18498, 12600, 18740,
+        12356, 12855, 20564, 12339, 18498, 12600, 18740, 12612, 14642, 20564,
+        12339, 18498, 12600, 18740, 12356, 13877, 0, 0, 0, 0, 0, 0, 0, 3843,
+        2306, 1282, 257, 0, 32340, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
+    ],
+    "Information": [12.0, 14, "H34***", 1, 1.15, 0.0, 1.14, 1.07, 0.0, 1]
+}
