+++ tests/test_base_inverter.py
@@ -1,6 +1,7 @@
 import pytest
 
 from solax import inverter
+from solax.discovery import REGISTRY
 
 
 @pytest.mark.asyncio
@@ -15,16 +16,6 @@ async def test_unimplemented_make_request_with_pwd():
         await inverter.Inverter.make_request('localhost', 80, 'pwd')
 
 
-def test_unimplemented_sensor_map():
-    with pytest.raises(NotImplementedError):
-        inverter.Inverter.sensor_map()
-
-
-def test_unimplemented_schema():
-    with pytest.raises(NotImplementedError):
-        inverter.Inverter.schema()
-
-
 def test_all_registered_inverters_inherit_from_base():
-    for i in inverter.REGISTRY:
+    for i in REGISTRY:
         assert issubclass(i, inverter.Inverter)
