+++ tests/test_discovery.py
@@ -1,7 +1,7 @@
 import pytest
 
 import solax
-from solax import inverter
+from solax.discovery import DiscoveryError
 
 
 @pytest.mark.asyncio
@@ -13,17 +13,17 @@ async def test_discovery(inverters_fixture):
 
 @pytest.mark.asyncio
 async def test_discovery_no_host():
-    with pytest.raises(inverter.DiscoveryError):
+    with pytest.raises(DiscoveryError):
         await solax.real_time_api('localhost', 2)
 
 
 @pytest.mark.asyncio
 async def test_discovery_no_host_with_pwd():
-    with pytest.raises(inverter.DiscoveryError):
+    with pytest.raises(DiscoveryError):
         await solax.real_time_api('localhost', 2, 'pwd')
 
 
 @pytest.mark.asyncio
 async def test_discovery_unknown_webserver(simple_http_fixture):
-    with pytest.raises(inverter.DiscoveryError):
+    with pytest.raises(DiscoveryError):
         await solax.real_time_api(*simple_http_fixture)
